% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LBPA_fits.R
\name{LBPA_fits}
\alias{LBPA_fits}
\title{Length-Based Pseudo-cohort Analysis - LBPA}
\usage{
LBPA_fits(name)
}
\arguments{
\item{name}{excel (.xlsx) file with data and parameters. This file contains four sheets: LF data, biological parameters,
initial parameters and data weighting}
}
\value{
Several graphics with de model fit and population variables are displayed. Four csv files are generated: model parameters,
population variables, log-likelihood and per-recruit variables

Several graphics with de model fit and population variables are displayed. Four csv files are generated: model parameters,
population variables, log-likelihood and per-recruit variables
}
\description{
The model estimates the fishing mortality and depletion of the Spawning Potential Ratio (SPR) in an equilibrium condition.
The model is fitted to several length frequencies of catches simultaneously by minimizing a penalized log-likelihood function.

The model estimates the fishing mortality and depletion of the Spawning Potential Ratio (SPR) in an equilibrium condition.
The model is fitted to several length frequencies of catches simultaneously by minimizing a penalized log-likelihood function.
}
\examples{
LBPA_fits("data_file.xlsx")
LBPA_fits("data_file.xlsx")
}
